Feature/command lists (#47)
* WIP CommandGroup & CommandList

* Implemented commandlist alloc/free functions

* CommandList begin/end implementation

* added unit testing structure for the newly introduced functions

* Renamed command_list_state::Executable to Ready and gave queue some helpful pointers

* Moved commandlist alloc/free functions to CommandGroup

* fixed destroyCommandGroup not taking nullptrs

* createCommandGroup() unit tests

* destroyCommandGroup unit tests

* added missing cleanup

* CommandGroup::reset unit tests

* Fixed incorrect result values

* 005_commands sample

* CommandList::begin() and CommandList::end() tests

* instance test bugfix

* unit tests cleanup

* WIP CommandGroup::allocate tests

* Resolves https://github.com/Legion-Engine/Legion-LLRI/pull/47#discussion_r698057851

* Resolves https://github.com/Legion-Engine/Legion-LLRI/pull/47#discussion_r698058766

* finished commandgroup unit tests

* commandgroup multiple recording commandlists validation

* removed VS files to simplify rebase

* added sample 005_commands to the cmake project

* fixed dx12 indirect commandlists

* added node masks to command list allocation

* updated the sample

* changed some tests to use nodemasks correctly

* added nodemask unit tests

* minor doc tweaks

* added Device::queryQueueCount() which returns the number of created queues

* fixed a memory leak in unit tests

* fixed incorrect queue counts in unit tests

* fixed vulkan multi-queue creation

* changed all for loops to use size_t instead of uintN_t

* moved findQueueFamilies to utils.hpp

* fixed cases where size_t shouldve stayed uint8_t

* few more fixes

+++ legion/engine/llri/detail/instance.inl
@@ -186,7 +186,8 @@ namespace LLRI_NAMESPACE
             { queue_type::Compute, 0 },
             { queue_type::Transfer, 0 }
         };
-        for (uint32_t i = 0; i < desc.numQueues; i++)
+
+        for (size_t i = 0; i < desc.numQueues; i++)
         {
             auto& queue = desc.queues[i];
